Bug #2030493 bsp: camera: camera provider 64bit
[root cause  ]: 64bit
[changes     ]:struct
[side effects]:no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums9620
[test case]:camera func
[reviewers]:yangyang.liu

commit_template_version:v1
[change_type ] feature_bugfix
[tag_product ] common

+++ core/cam_sys/inc/cam_buf.h
@@ -75,7 +75,7 @@ struct camera_buf {
 	uint32_t mfd;
 	struct dma_buf *dmabuf_p;
 	void *ionbuf;/* for iommu map */
-	uint32_t offset[3];
+	unsigned long offset[3];
 	size_t size;
 	unsigned long addr_vir[3];
 	unsigned long addr_k;
